+++ bin/abclinuxu_uploader.py
@@ -1,4 +1,4 @@
-#! /usr/bin/env python
+#! /usr/bin/env python2
 # -*- coding: utf-8 -*-
 #
 # Interpreter version: python 2.7
@@ -6,16 +6,14 @@
 # Imports =====================================================================
 import os
 import sys
+import shutil
 import os.path
 import getpass
 import argparse
 
 import dhtmlparser as d
 
-sys.path.insert(
-    0,
-    os.path.join(os.path.dirname(__file__), "../src")
-)
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "../src"))
 import abclinuxuapi
 
 
@@ -26,6 +24,7 @@ ALLOWED_IMAGES = [
     "gif",
     "png"
 ]
+MB = 1024 * 1024
 
 
 # Functions & objects =========================================================
@@ -37,17 +36,39 @@ def get_body(dom):
     return str(dom)
 
 
-def upload_image(concept, image_path):
+def upload_image(concept, image_path, images_url=None):
     # remote images
     if not os.path.exists(image_path):
         return image_path
+
+    if os.stat(image_path).st_size >= (MB - 1):
+        return _use_remote_url_for_image(image_path, images_url)
 
     concept.add_pic(open(image_path))
 
     return concept.list_pics()[-1]
 
 
-def upload_html(dom, args):
+def _use_remote_url_for_image(image_path, images_url):
+        if not images_url:
+            msg = "Image `%s` is bigger than 1 MB. Please supply --url parameter to link them (see --help).\n"
+            sys.stderr.write(msg % image_path)
+            sys.exit(1)
+
+        separator = "" if images_url.endswith("/") else "/"
+        url = images_url + separator + os.path.basename(image_path)
+
+        local_images_dir = "images_to_upload"
+        try:
+            os.makedirs(local_images_dir)
+        except Exception:
+            pass
+
+        shutil.copy(image_path, local_images_dir)
+        return url
+
+
+def upload_html(dom, args, images_url=None):
     user = abclinuxuapi.User(args.username, args.password)
 
     try:
@@ -69,7 +90,7 @@ def upload_html(dom, args):
 
         print "\tUploading '%s'" % os.path.basename(img.params["src"])
 
-        img.params["src"] = upload_image(concept, img.params["src"])
+        img.params["src"] = upload_image(concept, img.params["src"], images_url)
 
     print
     print "Uploading linked images:"
@@ -84,7 +105,7 @@ def upload_html(dom, args):
 
         print "\tUploading '%s'" % os.path.basename(a.params["href"])
 
-        a.params["href"] = upload_image(concept, a.params["href"])
+        a.params["href"] = upload_image(concept, a.params["href"], images_url)
 
     print
     print "Updating image links in concept .."
@@ -110,7 +131,7 @@ if __name__ == '__main__':
     )
     parser.add_argument(
         "-u",
-        '--username',
+        "--username",
         metavar="USER",
         required=True,
         type=str,
@@ -118,7 +139,7 @@ if __name__ == '__main__':
     )
     parser.add_argument(
         "-p",
-        '--password',
+        "--password",
         metavar="PASS",
         type=str,
         default=None,
@@ -129,8 +150,17 @@ if __name__ == '__main__':
         "--title",
         type=str,
         default=None,
-        help="Title of your blogpost. If not set, <title> tag from your\
+        help="Title of your blogpost. If not set, <title> tag from your \
               document is used."
+    )
+    parser.add_argument(
+        "--url",
+        type=str,
+        default=None,
+        help="URL where you'll put big images. At this moment, abclinuxu can't \
+              upload images bigger than 1 MB. All such files will be put into \
+              separate directory, which you can then upload to this URL and they \
+              will point there."
     )
     args = parser.parse_args()
 
@@ -138,7 +168,6 @@ if __name__ == '__main__':
         sys.stderr.write("File '%s' doesn't exists!\n" % args.FN)
         sys.exit(1)
 
-    dom = None
     with open(args.FN) as f:
         data = f.read()
         dom = d.parseString(data)
@@ -161,4 +190,4 @@ if __name__ == '__main__':
             os.path.abspath(args.FN)
         )
     )
-    upload_html(dom, args)
+    upload_html(dom, args, images_url=args.url)